[s23.org wikistats] Throttle updates for big farms, keep updating big wikis' stats
Patch quickly put together after my last comment on the bug.
If it looks reasonably ok, mutante will have to deploy it, change whatever switch
kept the wikia wikis stalled and then let the updates run to test.

Bug: 59943

PS2: sneak in a small live hack to keep it in sync with the labs instance

Change-Id: Ieb3c1d49b6c6b7208172f7e0bac810ce72fe3bc2