remove remnants of the 'views' column
The 'views' column doesn't really exist anymore in newer (WMF)-
Mediawikis and is about to be removed from core mw as well.

There also was the bug to
'completely remove or globally add the "views" column in stats tables'.

So this removes remnants of this old Views column in various place.
In all those places it wasn't used or was completely empty.
The only place it still appears is in "largest_csv" where some wikis
have a 'views' value.

Bug: T38293
Change-Id: Ic1958c8244534a9519334a81d72562bb91496106