build external links based on si_server,si_articlepath
This looks like a small change but it's actually quite big.

In the past we were using URLs with "Special:Statistics?action=raw"
and we were trying to guesstimate the links to other Special
pages by splitting URLs at the 'S' and replacing. But this was
broken for many, as evidenced by the request in T136183.

After this we are going to use the newer database fields
"si_server" and "si_articlepath" to build a more reliable URLs.

These come directly from the API(.php). It fixes it for the example
wiki lietuvai.lt which asked for the wiki and for many others.

It works for newer mediawikis that have these set and provide them
via API. Another prerequisite is that update.php ran with the "extinfo"
option for this wiki at least once, but that should be done for most where
it worked.

So this fixes it for many and it is still broken for others but it's definitely
the way to go and the reliable way to create URLs from a MW API.

Bug:T136183
Change-Id: I907235779a5b7758a4f2ff42c29ad2b9fbfa34a4