use https for wikis in $tables_with_suffix_wiki_last (w3c)
W3C wikis switched from http to https-only in 2016-05.

This broke updates (T149000), they all just got 301s.

We have a config option "$tables_https_only" but we don't
need to use that and add another check for that in the update.php
code since, for now, W3C wikis are the only ones with the
"$tables_with_suffix_wiki_last" URL format.

So we can just change the URL for those to use https, easier change.

Bug: T149000
Change-Id: I807bee25f5557eb7380454d479926f8cfe2c1bc6