miraheze auto import: change order, custom domains first
When this script ran and first imported the regular miraheze wikis,
then additionally added the custom domain wikis, the result was that
when custom wikis are imported they already exist with the same "prefix"
in the database (since custom wikis are ALSO appearing in the list of
regular wikis and not ONLY in the custom list, it seems as that last part
is something that might have changed on the Miraheze site recently to explain
why import first worked and then stopped working after a while).

Since the db column "prefix" is marked as unique the INSERT will be skipped
in this case. INSERT IGNORE does not change this, it just means the error
becomes a warning, but not that data gets inserted.

Anyways, the easiest fix here was to simply switch the order around,
import custom wikis first, then regular wikis, so if the custom ones get
skipped on import that is actually what we want and should happen anyways.

Bug: T158976
Change-Id: I746028126734143067f648d5951c409a6ba03995