remove all mysql_ functions, use PDO
The MySQL extension providing all the mysql_ functions
in PHP has been deprecated and is completely removed
in PHP7.

To migrate to stretch we have switch to mysqli or PDO,
and there are a lot of other advantages to it anyways.

Picking PDO because it supports client-side prepared statements
which make it much safer against SQL injection attacks.

Bug:T165684
Change-Id: Id748eea7b4ffcacc3e41627ba36a27006119df3a